modif

+++ src/app/layouts/header/header.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { ConfirmationComponent } from 'src/app/dialog/confirmation/confirmation.component';
+import swal from 'sweetalert2';
 
 @Component({
   selector: 'app-header',
@@ -17,16 +18,30 @@ export class HeaderComponent {
   }
 
   logout(){
-    const dialogConfig = new MatDialogConfig();
-    dialogConfig.data ={
-      message:'logout'
-    };
-    const dialogRef = this.dialog.open(ConfirmationComponent,dialogConfig);
-    const sub = dialogRef.componentInstance.onEmitStatusChange.subscribe((user)=>{
-      dialogRef.close();
-      localStorage.clear();
-      this.router.navigate(['/']);
-    })
+    // const dialogConfig = new MatDialogConfig();
+    // dialogConfig.data ={
+    //   message:'logout'
+    // };
+    // const dialogRef = this.dialog.open(ConfirmationComponent,dialogConfig);
+    // const sub = dialogRef.componentInstance.onEmitStatusChange.subscribe((user)=>{
+    //   dialogRef.close();
+    //   localStorage.clear();
+    //   this.router.navigate(['/']);
+    // })
+     
+swal.fire({
+  title: 'Voulez-vous quitter l\'application?',
+  showCancelButton: true,
+  confirmButtonText: 'Yes',
+  confirmButtonColor: '#3085d6',
+  cancelButtonColor: '#d33',
+}).then((result) => {
+  /* Read more about isConfirmed, isDenied below */
+  if (result.isConfirmed) {
+    localStorage.clear();
+    this.router.navigate(['/']);
+  } 
+})
   }
 
   changePassword(){
